Add a test suite testing the seal fields decode functions
